Move Uniswap pool fetching logic into shared component
