+++ shared/Cargo.toml
@@ -5,7 +5,12 @@ edition = "2018"
 
 [dependencies]
 atty = "0.2"
+contracts = { path = "../contracts" }
+ethcontract = { version = "0.11", default-features = false }
+hex-literal = "0.3"
+model = { path = "../model" }
 structopt = { version = "0.3", default-features = false }
 tracing = "0.1"
 tracing-subscriber = "0.2"
 url = "2.2"
+web3 = { version = "0.15", default-features = false }
