+++ shared/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod arguments;
 pub mod time;
 pub mod tracing;
+pub mod uniswap_pool;
